refactored all classes in order to store double as data, soon it's better to use templates

+++ include/LinearAlgebra.hpp
@@ -10,7 +10,7 @@ namespace LinearAlgebra{
     {
         unsigned _len;
         
-        int*     _vec;
+        double*  _vec;
 
         public:
             Vector(unsigned len);
@@ -18,27 +18,27 @@ namespace LinearAlgebra{
             Vector(Vector&& v);
             ~Vector();
 
-            void randomInit(int a,int b);
-            void valInit(int val);
+            void randomInit(double a,double b);
+            void valInit(double val);
 
             Vector operator+(const Vector& other)const;
             Vector operator-(const Vector& other)const;
             Vector operator*(const Vector& other)const;
             Vector operator/(const Vector& other)const;
 
-            Vector operator+(const int constant)const;
-            Vector operator-(const int constant)const;
-            Vector operator*(const int constant)const;
-            Vector operator/(const int constant)const;
+            Vector operator+(const double constant)const;
+            Vector operator-(const double constant)const;
+            Vector operator*(const double constant)const;
+            Vector operator/(const double constant)const;
 
             Vector gpu_diff(const Vector& v2)const;
             Vector gpu_sum(const Vector& v2)const;
 
             unsigned len()const;
-            int*     getVec();
+            double*  getVec();
 
-            int& operator [](unsigned i);
-            const int& operator [](unsigned i)const;
+            double& operator [](unsigned i);
+            const double& operator [](unsigned i)const;
             bool operator==(const Vector& other) const;
 
             friend std::ostream& operator<<(std::ostream& stream, const Vector& operand);
@@ -49,7 +49,7 @@ namespace LinearAlgebra{
         unsigned _rows;
         unsigned _cols;
 
-        int* _data;
+        double*  _data;
 
         public:
             Matrix(unsigned rows, unsigned cols);
@@ -57,18 +57,18 @@ namespace LinearAlgebra{
             Matrix(Matrix&& mat);
             ~Matrix();
 
-            void randomInit(int a,int b);
-            void valInit(int val);
+            void randomInit(double a,double b);
+            void valInit(double val);
 
             Matrix operator+(const Matrix& other)const;
             Matrix operator-(const Matrix& other)const;
             Matrix operator*(const Matrix& other)const;
             Matrix operator/(const Matrix& other)const;
 
-            Matrix operator+(const int constant)const;
-            Matrix operator-(const int constant)const;
-            Matrix operator*(const int constant)const;
-            Matrix operator/(const int constant)const;
+            Matrix operator+(const double constant)const;
+            Matrix operator-(const double constant)const;
+            Matrix operator*(const double constant)const;
+            Matrix operator/(const double constant)const;
 
             Vector matrixVectorMult(const Vector& v1)const;
             Matrix matrixMultiplication(const Matrix& mat)const;
@@ -81,10 +81,10 @@ namespace LinearAlgebra{
 
             unsigned rows()const;
             unsigned cols()const;
-            int*     data();
+            double*  data();
 
-            int& operator [](unsigned i);
-            const int& operator [](unsigned i)const;
+            double& operator [](unsigned i);
+            const double& operator [](unsigned i)const;
             bool operator==(const Matrix& other) const;
 
             friend std::ostream& operator<<(std::ostream& stream, const Matrix& operand);
@@ -98,7 +98,7 @@ namespace LinearAlgebra{
 
         unsigned *_rows;
         unsigned *_cols;
-        int      *_vals;
+        double   *_vals;
 
         public:
             CSRMatrix(unsigned nRows,unsigned nCols, unsigned nNzElems);
@@ -106,7 +106,7 @@ namespace LinearAlgebra{
             CSRMatrix(CSRMatrix&& mat);
             ~CSRMatrix();
 
-            void randomInit(int a,int b);
+            void randomInit(double a,double b);
 
             Vector matrixVectorMult(const Vector& v1)const;
 
@@ -120,7 +120,7 @@ namespace LinearAlgebra{
             unsigned  nonZeroElements()const;
             unsigned* getRowsArray();
             unsigned* getColsArray();
-            int*      getValsArray();
+            double*   getValsArray();
 
             bool operator==(const CSRMatrix& other) const;
 
@@ -135,7 +135,7 @@ namespace LinearAlgebra{
 
         unsigned *_cols;
         unsigned *_rows;
-        int      *_vals;
+        double   *_vals;
 
         public:
             CSCMatrix(unsigned nRows,unsigned nCols, unsigned nNzElems);
@@ -143,7 +143,7 @@ namespace LinearAlgebra{
             CSCMatrix(CSCMatrix&& mat);
             ~CSCMatrix();
 
-            void randomInit(int a,int b);
+            void randomInit(double a,double b);
 
             Vector matrixVectorMult(const Vector& v1)const;
 
@@ -157,7 +157,7 @@ namespace LinearAlgebra{
             unsigned  nonZeroElements()const;
             unsigned* getColsArray();
             unsigned* getRowsArray();
-            int*      getValsArray();
+            double*   getValsArray();
 
             bool operator==(const CSCMatrix& other) const;
