+++ src/MatrixKernels.cu
@@ -1,10 +1,10 @@
-__global__ void matrixVectorMultKernel(const int* matrix, const int* v1, int* rv, const unsigned rows, const unsigned cols)
+__global__ void matrixVectorMultKernel(const double* matrix, const double* v1, double* rv, const unsigned rows, const unsigned cols)
 {
     const unsigned row = blockIdx.x * blockDim.x + threadIdx.x;
 
     if(row >= rows) return;
 
-    int acc = 0;
+    double acc = 0;
 
     for(unsigned i = 0u; i < cols; i++ )
     {
