+++ src/VectorKernels.cu
@@ -1,4 +1,4 @@
-__global__ void vectorDifKernel(const int* v1, const int* v2, int* rv, const unsigned size)
+__global__ void vectorDifKernel(const double* v1, const double* v2, double* rv, const unsigned size)
 {
     const unsigned globalIndex = blockIdx.x * blockDim.x + threadIdx.x;
